Implemented data table for shipment block. Refactoring & cleanup.
